+++ src/femto/helpers.py
@@ -4,10 +4,7 @@ import functools
 import itertools
 import os
 import pathlib
-from typing import Any
-from typing import Generator
-from typing import Iterable
-from typing import Iterator
+from typing import Any, Generator, Iterable, Iterator
 
 import numpy as np
 import numpy.typing as npt
@@ -205,7 +202,7 @@ def sign() -> Iterator[int]:
     return itertools.cycle([1, -1])
 
 
-def remove_repeated_coordinates(array: npt.NDArray[np.float32]) -> npt.NDArray[np.float32]:
+def remove_repeated_coordinates(array: npt.NDArray[np.float64]) -> npt.NDArray[np.float64]:
     """Remove repeated coordinates.
 
     Parameters
@@ -222,11 +219,11 @@ def remove_repeated_coordinates(array: npt.NDArray[np.float32]) -> npt.NDArray[n
 
     mask = np.diff(array).astype(bool)
     mask = np.insert(mask, 0, True)
-    return np.where(~mask, np.nan, array).astype(np.float32)
+    return np.where(~mask, np.nan, array).astype(np.float64)
 
 
 # Filtering adjacent identical points from a list of arrays.
-def unique_filter(arrays: list[npt.NDArray[np.float32]]) -> npt.NDArray[np.float32]:
+def unique_filter(arrays: list[npt.NDArray[np.float64]]) -> npt.NDArray[np.float64]:
     """Remove duplicate subsequent points.
 
     Filtering adjacent identical points from a list of arrays.
@@ -278,7 +275,7 @@ def unique_filter(arrays: list[npt.NDArray[np.float32]]) -> npt.NDArray[np.float
     if len(arrays) == 1:
         data = arrays[0]
     else:
-        data = np.stack(arrays, axis=-1).astype(np.float32)
+        data = np.stack(arrays, axis=-1).astype(np.float64)
     data.reshape(-1, len(arrays))
 
     # mask
